add 32-bit support and fix residuals bug

+++ include/rt-include.h
@@ -212,11 +212,11 @@ struct residual_pointer_t {
     uint32_t      dst_sig  ;
     uint32_t      free_sig ;
     uint32_t      store_id ;
-    int64_t       counter  ;
-    int64_t       adj_cnt  ;
+    int32_t       counter  ;
+    int32_t       adj_cnt  ;
 
     residual_pointer_t (uintptr_t l, uintptr_t v, uint32_t s, uint32_t d,
-            uint32_t f, uint32_t id, int64_t c) {
+            uint32_t f, uint32_t id, int32_t c) {
         loc = l;
         val = v;
         src_sig = s;
